+++ subgraph/tests/keys-v11.test.ts
@@ -27,10 +27,8 @@ import {
   createKeyExtendedEvent,
   createKeyManagerChangedEvent,
   createRenewKeyPurchaseEvent,
-  mockDataSourceV8,
   mockDataSourceV11,
   updateExpiration,
-  updateExpirationV8,
 } from './keys-utils'
 import {
   keyOwnerAddress,
@@ -40,7 +38,6 @@ import {
   tokenURI,
   expiration,
   lockAddress,
-  lockAddressV8,
   lockManagers,
 } from './constants'
 
@@ -48,34 +45,6 @@ import {
 import './mocks'
 
 const keyID = `${lockAddress}-${tokenId}`
-const keyIDV8 = `${lockAddressV8}-${tokenId}`
-
-describe('Key transfers (v8)', () => {
-  test('Creation of a new key (v8)', () => {
-    mockDataSourceV8()
-    const newTransferEvent = createTransferEvent(
-      Address.fromString(nullAddress),
-      Address.fromString(keyOwnerAddress),
-      BigInt.fromU32(tokenId)
-    )
-    handleTransfer(newTransferEvent)
-
-    assert.entityCount('Key', 1)
-    assert.fieldEquals('Key', keyIDV8, 'lock', lockAddressV8)
-    assert.fieldEquals('Key', keyIDV8, 'owner', keyOwnerAddress)
-    assert.fieldEquals('Key', keyIDV8, 'tokenId', `${tokenId}`)
-    assert.fieldEquals('Key', keyIDV8, 'tokenURI', `${tokenURI}`)
-    assert.fieldEquals('Key', keyIDV8, 'expiration', `${expiration}`)
-    assert.fieldEquals('Key', keyIDV8, 'createdAtBlock', '1')
-    assert.fieldEquals('Key', keyIDV8, 'createdAt', '1')
-    assert.fieldEquals('Key', keyIDV8, 'manager', lockManagers[0])
-  })
-
-  afterAll(() => {
-    clearStore()
-    dataSourceMock.resetValues()
-  })
-})
 
 describe('Burn a key', () => {
   beforeAll(() => {
@@ -189,29 +158,6 @@ describe('Change in expiration timestamp', () => {
     assert.fieldEquals('Key', keyID, 'expiration', `${expiration + 1000}`)
     dataSourceMock.resetValues()
   })
-
-  test('should increase key timestamp (v8)', () => {
-    mockDataSourceV8()
-    // create a key
-    const newTransferEvent = createTransferEvent(
-      Address.fromString(nullAddress),
-      Address.fromString(keyOwnerAddress),
-      BigInt.fromU32(tokenId)
-    )
-    handleTransfer(newTransferEvent)
-
-    // mock and test
-    updateExpirationV8(BigInt.fromU64(expiration + 1000))
-    const newExpirationEvent = createExpirationChangedEventUntilV11(
-      BigInt.fromU32(tokenId),
-      BigInt.fromU32(1000),
-      true
-    )
-
-    handleExpirationChangedUntilV11(newExpirationEvent)
-    assert.fieldEquals('Key', keyIDV8, 'expiration', `${expiration + 1000}`)
-    dataSourceMock.resetValues()
-  })
 })
 
 describe('Extend key', () => {
@@ -270,25 +216,6 @@ describe('Key is expired by lock manager', () => {
     assert.fieldEquals('Key', keyID, 'expiration', `${now}`)
     dataSourceMock.resetValues()
   })
-
-  test('should update the key expiration (v8)', () => {
-    mockDataSourceV8()
-    // create a key
-    const newTransferEvent = createTransferEvent(
-      Address.fromString(nullAddress),
-      Address.fromString(keyOwnerAddress),
-      BigInt.fromU32(tokenId)
-    )
-    handleTransfer(newTransferEvent)
-
-    // mock and test
-    updateExpirationV8()
-    const newExpireKeyEvent = createExpireKeyEvent(BigInt.fromU32(tokenId))
-
-    handleExpireKey(newExpireKeyEvent)
-    assert.fieldEquals('Key', keyIDV8, 'expiration', `${now}`)
-    dataSourceMock.resetValues()
-  })
 })
 
 describe('Key managers', () => {
@@ -332,27 +259,6 @@ describe('Key managers', () => {
     assert.fieldEquals('Key', keyID, 'manager', newKeyManagerAddress)
     dataSourceMock.resetValues()
   })
-
-  test('key manager changed (v8)', () => {
-    mockDataSourceV8()
-
-    // create a key
-    const newTransferEvent = createTransferEvent(
-      Address.fromString(nullAddress),
-      Address.fromString(keyOwnerAddress),
-      BigInt.fromU32(tokenId)
-    )
-    handleTransfer(newTransferEvent)
-
-    const newKeyManagerChanged = createKeyManagerChangedEvent(
-      BigInt.fromU32(tokenId),
-      Address.fromString(newKeyManagerAddress)
-    )
-
-    handleKeyManagerChanged(newKeyManagerChanged)
-    assert.fieldEquals('Key', keyIDV8, 'manager', newKeyManagerAddress)
-    dataSourceMock.resetValues()
-  })
 })
 
 describe('Cancel keys', () => {
@@ -394,29 +300,7 @@ describe('Cancel keys', () => {
   })
 })
 
-describe('RenewKeyPurchase (lock <v10)', () => {
-  test('extend a key by the correct time (v8)', () => {
-    mockDataSourceV8()
-
-    // create a key
-    const newTransferEvent = createTransferEvent(
-      Address.fromString(nullAddress),
-      Address.fromString(keyOwnerAddress),
-      BigInt.fromU32(tokenId)
-    )
-    handleTransfer(newTransferEvent)
-
-    updateExpirationV8()
-    const newExpiration = expiration + 1000
-    const newRenewKeyPurchase = createRenewKeyPurchaseEvent(
-      Address.fromString(keyOwnerAddress),
-      BigInt.fromU64(newExpiration)
-    )
-    handleRenewKeyPurchase(newRenewKeyPurchase)
-    assert.fieldEquals('Key', keyIDV8, 'expiration', `${newExpiration}`)
-    dataSourceMock.resetValues()
-  })
-
+describe('RenewKeyPurchase', () => {
   test('should add the new transaction to the list of transactions', () => {
     mockDataSourceV11()
     const newTransferEvent = createTransferEvent(
